analytics/satellite_work/requirements.txt
aenum==3.1.5
affine==2.3.0
appdirs==1.4.4
attrs==21.2.0
autopep8==1.6.0
bleach==4.1.0
bokeh==2.4.2
boto3==1.20.18
botocore==1.23.18
Cartopy @ file:///C:/Repos/open_hack_data/Cartopy-0.20.1-cp38-cp38-win_amd64.whl
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
colorcet==3.0.0
cycler==0.11.0
dataclasses-json==0.5.6
eo-learn-core==0.10.1
eo-learn-io==0.10.1
eo-learn-visualization[FULL]
Fiona @ file:///C:/Repos/open_hack_data/Fiona-1.8.20-cp38-cp38-win_amd64.whl
fonttools==4.28.2
fs==2.4.14
fs-s3fs==1.1.1
GDAL @ file:///C:/Repos/open_hack_data/GDAL-3.3.3-cp38-cp38-win_amd64.whl
geopandas==0.10.2
geoviews==1.9.2
graphviz==0.19
holoviews==1.14.6
hvplot==0.7.3
idna==3.3
importlib-metadata==4.8.2
Jinja2==3.0.3
jmespath==0.10.0
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib==3.5.0
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
packaging==21.3
pandas==1.3.4
panel==0.12.5
param==1.12.0
Pillow==8.4.0
pycodestyle==2.8.0
pyct==0.4.8
pydot==1.4.2
pyepsg==0.4.0
Pygments==2.10.0
pyparsing==3.0.6
pyproj==3.3.0
pyshp==2.1.3
python-dateutil==2.8.2
pytz==2021.3
pyviz-comms==2.1.0
PyYAML==6.0
rasterio @ file:///C:/Repos/open_hack_data/rasterio-1.2.10-cp38-cp38-win_amd64.whl
requests==2.26.0
requests-oauthlib==1.3.0
Rtree==0.9.7
s3transfer==0.5.0
sentinelhub==3.4.1
setuptools-scm==6.3.2
Shapely==1.8.0
six==1.16.0
snuggs==1.4.7
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
utm==0.7.0
webencodings==0.5.1
xarray==0.20.1
zipp==3.6.0
